Added db connections

+++ bookstore_backend/bookstore/src/main/java/com/listek/bookstore/repositories/UserAccountRepository.java
@@ -3,5 +3,5 @@ package com.listek.bookstore.repositories;
 import com.listek.bookstore.models.UserAccount;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserAccountRepository extends CrudRepository<UserAccount, Long> {
+public interface UserAccountRepository extends CrudRepository<UserAccount, Integer> {
 }
